requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn>=0.30.1,<1.0.0
python-dotenv>=1.0.1,<2.0.0
environs>=11.0.0,<12.0.0
aiohttp>=3.9.5,<4.0.0
asyncpg>=0.29.0,<1.0.0
SQLAlchemy[asyncio]>=2.0.30,<3.0.0
pgvector>=0.2.5,<0.3.0
openai>=1.34.0,<2.0.0
tiktoken>=0.7.0,<0.8.0
openai-messages-token-helper>=0.1.8,<0.2.0
rich>=13.0.0,<14.0.0
fastapi==0.115.2
uvicorn==0.31.1